// libs/express/api/cache/src/lib/types.ts
export interface ICache {
  getAsync: (key: string) => Promise<string | null>;
  setAsync: (
    key: string,
    value: string,
    mode?: SetAsyncMode,
    duration?: number
  ) => Promise<unknown>;
  delAsync: (key: string) => Promise<number>;
  expireAsync: (key: string, second: number) => Promise<number>;
  incrByAsync: (key: string, increment: number) => Promise<number>;
  decrByAsync: (key: string, decrement: number) => Promise<number>;
}

export type Config = {
  host?: string;
  port?: number;
  password?: string;
};

export type SetAsyncMode = 'EX' | 'PX' | 'KEEPTTL';